feat: add timer

+++ task3/src/main/java/ru/cft/miner/model/observer/TimerListener.java
@@ -2,5 +2,5 @@ package ru.cft.miner.model.observer;
 
 public interface TimerListener {
 
-    void onTimeChanged(long time);
+    void onTimeChanged(int time);
 }
